+++ UI/imgui.ini
@@ -9,8 +9,8 @@ Size=848,709
 Collapsed=0
 
 [Window][Hello, world!]
-Pos=318,698
-Size=1215,319
+Pos=317,816
+Size=1216,201
 Collapsed=0
 DockId=0x00000021,0
 
@@ -21,7 +21,7 @@ Collapsed=0
 
 [Window][explorer]
 Pos=0,19
-Size=316,998
+Size=315,998
 Collapsed=0
 DockId=0x00000001,0
 
@@ -62,16 +62,16 @@ Size=414,335
 Collapsed=0
 
 [Window][viewport #1##viewport_window]
-Pos=318,19
-Size=1215,677
+Pos=317,19
+Size=1216,795
 Collapsed=0
-DockId=0x00000013,0
+DockId=0x0000000B,0
 
 [Window][viewport #2##viewport_window]
-Pos=912,19
-Size=600,828
+Pos=317,419
+Size=1216,395
 Collapsed=0
-DockId=0x0000001C,0
+DockId=0x0000000C,0
 
 [Window][viewport #3##viewport_window]
 Pos=311,114
@@ -96,8 +96,8 @@ Size=270,123
 Collapsed=0
 
 [Window][new cylinder##new_cylinder_modal_window]
-Pos=827,458
-Size=266,100
+Pos=828,458
+Size=263,100
 Collapsed=0
 
 [Window][new plane##new_plane_modal_window]
@@ -260,7 +260,7 @@ DockId=0xA086D808,4
 Pos=917,756
 Size=599,261
 Collapsed=0
-DockId=0x00000013,0
+DockId=0x0000000B,0
 
 [Window][settings]
 Pos=318,698
@@ -381,9 +381,9 @@ Column 1  Weight=1.0000
 [Docking][Data]
 DockSpace                             ID=0x8B93E3BD Window=0xA787BDB4 Pos=0,42 Size=1920,998 Split=X Selected=0xE87781F4
   DockNode                            ID=0x00000005 Parent=0x8B93E3BD SizeRef=1533,998 Split=X
-    DockNode                          ID=0x00000001 Parent=0x00000005 SizeRef=316,998 Selected=0xC24E1FFE
-    DockNode                          ID=0x00000002 Parent=0x00000005 SizeRef=1215,998 Split=Y
-      DockNode                        ID=0x0000001D Parent=0x00000002 SizeRef=1285,322 Split=X
+    DockNode                          ID=0x00000001 Parent=0x00000005 SizeRef=315,998 Selected=0xC24E1FFE
+    DockNode                          ID=0x00000002 Parent=0x00000005 SizeRef=1216,998 Split=Y
+      DockNode                        ID=0x0000001D Parent=0x00000002 SizeRef=1285,498 Split=X
         DockNode                      ID=0x00000009 Parent=0x0000001D SizeRef=1571,1017 Split=X
           DockNode                    ID=0x00000007 Parent=0x00000009 SizeRef=1508,1017 Split=X
             DockNode                  ID=0x00000003 Parent=0x00000007 SizeRef=1178,1017 Split=X
@@ -393,9 +393,9 @@ DockSpace                             ID=0x8B93E3BD Window=0xA787BDB4 Pos=0,42 S
                     DockNode          ID=0x00000017 Parent=0x00000011 SizeRef=603,798 Split=X Selected=0xF0955E0B
                       DockNode        ID=0x00000015 Parent=0x00000017 SizeRef=0,0 Split=X Selected=0x63715C1D
                         DockNode      ID=0x00000014 Parent=0x00000015 SizeRef=571,694 Split=Y Selected=0xB50D76B1
-                          DockNode    ID=0x0000001A Parent=0x00000014 SizeRef=514,390 Split=X Selected=0xB50D76B1
-                            DockNode  ID=0x00000013 Parent=0x0000001A SizeRef=599,828 CentralNode=1 Selected=0xB50D76B1
-                            DockNode  ID=0x0000001C Parent=0x0000001A SizeRef=600,828 Selected=0x6D112248
+                          DockNode    ID=0x0000001A Parent=0x00000014 SizeRef=514,390 Split=Y Selected=0xB50D76B1
+                            DockNode  ID=0x0000000B Parent=0x0000001A SizeRef=1216,398 CentralNode=1 Selected=0xB50D76B1
+                            DockNode  ID=0x0000000C Parent=0x0000001A SizeRef=1216,395 Selected=0x6D112248
                           DockNode    ID=0x0000001B Parent=0x00000014 SizeRef=514,554 Selected=0x9B5DBF31
                         DockNode      ID=0x00000019 Parent=0x00000015 SizeRef=669,694 Selected=0x8C75DBF1
                       DockNode        ID=0x00000016 Parent=0x00000017 SizeRef=619,798 Selected=0xDFFE6FBA
@@ -406,7 +406,7 @@ DockSpace                             ID=0x8B93E3BD Window=0xA787BDB4 Pos=0,42 S
             DockNode                  ID=0x00000004 Parent=0x00000007 SizeRef=740,1017 Selected=0x06B5FAA8
           DockNode                    ID=0x00000008 Parent=0x00000009 SizeRef=410,1017 Selected=0x250DD6E2
         DockNode                      ID=0x0000000A Parent=0x0000001D SizeRef=347,1017 Selected=0x62AEB663
-      DockNode                        ID=0x0000001E Parent=0x00000002 SizeRef=1285,319 Split=X Selected=0xD00C9CD4
+      DockNode                        ID=0x0000001E Parent=0x00000002 SizeRef=1285,201 Split=X Selected=0xD00C9CD4
         DockNode                      ID=0x00000020 Parent=0x0000001E SizeRef=334,319 Selected=0x4C43E9BB
         DockNode                      ID=0x00000021 Parent=0x0000001E SizeRef=895,319 Selected=0xD00C9CD4
   DockNode                            ID=0x00000006 Parent=0x8B93E3BD SizeRef=385,998 Selected=0xD4FECC4E
