+++ UI/properties.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "rayzath.h"
 #include "explorer_base.hpp"
+#include "rayzath.h"
 
 #include "imgui.h"
 
@@ -11,80 +11,6 @@ namespace RayZath::UI::Windows
 {
 	namespace RZ = RayZath::Engine;
 	using ObjectType = RZ::World::ObjectType;
-
-	template <ObjectType T>
-	class Search
-	{
-	public:
-		using object_t = RZ::World::object_t<T>;
-	private:
-		bool m_opened = true;
-		std::array<char, 256> m_buffer{};
-		bool m_match_case = false, m_match_word = false;
-
-		RZ::Handle<object_t> m_selected;
-
-	private:
-		bool matches(const std::string& name)
-		{
-			std::string search_input(m_buffer.data());
-
-			static_assert(std::tuple_size_v<decltype(m_buffer)> > 0);
-			if (m_buffer[0] == '\0') return true; // no input, matches all
-
-			if (m_match_case)
-			{
-				return name.find(search_input) != std::string::npos;
-			}
-			else
-			{
-				std::transform(search_input.begin(), search_input.end(), search_input.begin(),
-					[](auto c) { return std::tolower(c); });
-				std::string lowered_name = name;
-				std::transform(lowered_name.begin(), lowered_name.end(), lowered_name.begin(),
-					[](auto c) { return std::tolower(c); });
-				return lowered_name.find(search_input) != std::string::npos;
-			}
-		}
-	public:
-		std::tuple<bool, RZ::Handle<object_t>> update(const RZ::World& world)
-		{
-			static constexpr auto* popup_id = "search##search_popup_modal";
-			ImGui::OpenPopup(popup_id);
-			if (ImGui::BeginPopupModal(popup_id, &m_opened))
-			{
-				ImGui::SetNextItemWidth(-1);
-				ImGui::InputTextWithHint("##object_name_input", "name",
-					m_buffer.data(), m_buffer.size(),
-					ImGuiInputTextFlags_AlwaysOverwrite);
-				ImGui::Checkbox("match case", &m_match_case);
-
-				ImGui::Separator();
-
-				if (ImGui::BeginListBox("search_list_box", ImVec2(-1.f, -1.f)))
-				{
-					auto& objects = world.Container<T>();
-					for (uint32_t i = 0; i < objects.GetCount(); i++)
-					{
-						const auto& object = objects[i];
-						if (!object) continue;
-						if (!matches(object->GetName())) continue;
-
-						bool is_selected = m_selected == object;
-						if (ImGui::Selectable(object->GetName().c_str(), &is_selected))
-							m_selected = object;
-
-						if (is_selected)
-							ImGui::SetItemDefaultFocus();
-					}
-					ImGui::EndListBox();
-				}
-				ImGui::EndPopup();
-			}
-
-			return std::make_tuple(m_opened, m_selected);
-		}
-	};
 
 	template <Engine::World::ObjectType T>
 	class PropertiesBase
@@ -109,7 +35,7 @@ namespace RayZath::UI::Windows
 	class Properties;
 
 	template<>
-	class Properties<Engine::World::ObjectType::Material> 
+	class Properties<Engine::World::ObjectType::Material>
 		: public PropertiesBase<Engine::World::ObjectType::Material>
 		, public ExplorerSelectable
 	{
@@ -134,7 +60,7 @@ namespace RayZath::UI::Windows
 		void display(RZ::Material& material);
 	};
 	template<>
-	class Properties<Engine::World::ObjectType::MeshStructure> 
+	class Properties<Engine::World::ObjectType::MeshStructure>
 		: public PropertiesBase<Engine::World::ObjectType::MeshStructure>
 	{
 		using PropertiesBase<Engine::World::ObjectType::MeshStructure>::m_object;
@@ -169,7 +95,7 @@ namespace RayZath::UI::Windows
 		void display();
 	};
 	template<>
-	class Properties<Engine::World::ObjectType::Mesh> 
+	class Properties<Engine::World::ObjectType::Mesh>
 		: public PropertiesBase<Engine::World::ObjectType::Mesh>
 		, public ExplorerSelectable
 	{
